Refactoring
- Fixed facebook auto-login
