+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
         </activity>
 
         <receiver
-            android:name=".GCMBroadcastReceiver"
+            android:name=".gcm.GCMBroadcastReceiver"
             android:permission="com.google.android.c2dm.permission.SEND" >
 
             <intent-filter>
@@ -42,7 +42,7 @@
             </intent-filter>
         </receiver>
 
-        <service android:name=".GcmIntentService" />
+        <service android:name=".gcm.GCMIntentService" />
 
         <activity android:name="com.facebook.LoginActivity"
             android:theme="@android:style/Theme.Translucent.NoTitleBar"
