+++ app/src/main/java/io/scrollback/app/gcm/GCMBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package io.scrollback.app;
+package io.scrollback.app.gcm;
 
 import android.app.Activity;
 import android.content.ComponentName;
@@ -13,7 +13,7 @@ public class GCMBroadcastReceiver extends WakefulBroadcastReceiver {
     public void onReceive(Context context, Intent intent) {
         // Explicitly specify that GcmIntentService will handle the intent.
         ComponentName comp = new ComponentName(context.getPackageName(),
-                GcmIntentService.class.getName());
+                GCMIntentService.class.getName());
         // Start the service, keeping the device awake while it is launching.
         startWakefulService(context, (intent.setComponent(comp)));
         setResultCode(Activity.RESULT_OK);
